- modal for deleting comment

+++ src/app/+shared/components/comment-box/comment-box.component.ts
@@ -106,6 +106,19 @@ export class CommentBoxComponent implements OnInit {
   }
 
   removeComment(id: string) {
-    this.commentService.deleteComment(id);
+
+      const dialogRef = this.dialog.open(AlertComponent, {
+        width: '300px',
+        data: {
+          message: 'Are you sure you want to delete this comment?',
+          type: 'confirmDelete'
+        }
+      });
+
+      dialogRef.afterClosed().subscribe(result => {
+        if (result && result.response) {
+          this.commentService.deleteComment(id);
+        }
+      });
   }
 }
